+++ ECHOTests/Stubs/JSONCodingStubs/FullAccountStub.swift
@@ -35,11 +35,8 @@ struct FullAccountStub {
                                 },
                                 "echorand_key": "fd9370d6384ee109765616b2ec5337b221a53a1fa3f3a076b68fcf774994a914",
                                 "options": {
-                                    "voting_account": "1.2.5",
                                     "delegating_account": "1.2.12",
                                     "num_witness": 0,
-                                    "num_committee": 0,
-                                    "votes": [],
                                     "extensions": []
                                 },
                                 "statistics": "2.6.22",
@@ -70,7 +67,6 @@ struct FullAccountStub {
                             "registrar_name": "nathan",
                             "referrer_name": "nathan",
                             "lifetime_referrer_name": "nathan",
-                            "votes": [],
                             "balances": [
                                 {
                                     "id": "2.5.11",