+++ ECHOTests/Stubs/JSONCodingStubs/OptionsStub.swift
@@ -9,13 +9,8 @@
 struct OptionsStub {
     let initialJson = """
                                 {
-                        "voting_account": "1.2.5",
                         "delegating_account": "1.2.12",
                         "num_witness": 0,
-                        "num_committee": 0,
-                        "votes": [
-                          
-                        ],
                         "extensions": [
                           
                         ]